Add all pending changes and updates

+++ Repositories/wishlist/WishlistRepository.cs
@@ -4,7 +4,7 @@ using PetGrubBakcend.Entities;
 
 namespace PetGrubBakcend.Repositories.wishlist
 {
-    public class WishlistRepository:IWishlistRepository
+    public class WishlistRepository : IWishlistRepository
     {
         private readonly AppDbContext _context;
         public WishlistRepository(AppDbContext context)
@@ -15,7 +15,7 @@ namespace PetGrubBakcend.Repositories.wishlist
 
         public async Task<bool> ExistAsync(int wishlistId, int productId)
         {
-          return await _context.Wishlist.AnyAsync(w => w.Id == wishlistId && w.ProductId == productId);
+            return await _context.Wishlist.AnyAsync(w => w.Id == wishlistId && w.ProductId == productId);
         }
 
         public async Task AddAsync(Wishlist wishlist)
@@ -25,7 +25,7 @@ namespace PetGrubBakcend.Repositories.wishlist
                 _context.Add(wishlist);
                 await _context.SaveChangesAsync();
             }
-            catch(Exception ex)
+            catch (Exception ex)
             {
                 throw;
             }
@@ -36,11 +36,11 @@ namespace PetGrubBakcend.Repositories.wishlist
         {
             try
             {
-               var res = await _context.Wishlist
-                    .Include(w => w.Product)
-                    .Where(w => w.UserId == userId)
-                    .Select(w => w.Product)
-                    .ToListAsync();
+                var res = await _context.Wishlist
+                     .Include(w => w.Product)
+                     .Where(w => w.UserId == userId)
+                     .Select(w => w.Product)
+                     .ToListAsync();
                 return res;
             }
             catch
@@ -49,16 +49,15 @@ namespace PetGrubBakcend.Repositories.wishlist
             }
         }
 
-        public async Task RemoveWishlisted(int userId,int productId)
+        public async Task RemoveWishlisted(int userId, int productId)
         {
-          var existing = await _context.Wishlist.FirstOrDefaultAsync(w => userId == w.UserId && productId == w.ProductId);
-          
+            var existing = await _context.Wishlist.FirstOrDefaultAsync(w => userId == w.UserId && productId == w.ProductId);
+
+            if (existing != null)
+            {
                 _context.Wishlist.Remove(existing);
                 await _context.SaveChangesAsync();
-         
+            }
         }
-
-
-
     }
 }
